requirements.txt
python-telegram-bot==13.7
gspread==3.7.0
oauth2client==4.1.3
python-dotenv==0.19.0
black==23.7.0